// packages/wcl/src/types.ts
export type InputMaybe<T> = Maybe<T>;
// ...
export type GetFightsQueryVariables = Exact<{
  reportID: Scalars["String"]["input"]
}>;

export interface GetFightsQuery {
  __typename?: "Query"
  reportData?: {
    __typename?: "ReportData"
    report?: {
      __typename?: "Report"
      title: string
      startTime: number
      endTime: number
      region?: { __typename?: "Region", slug: string } | null
      fights?: Array<{
        __typename?: "ReportFight"
        id: number
        startTime: number
        endTime: number
        difficulty?: number | null
        encounterID: number
        friendlyPlayers?: Array<number | null> | null
      } | null> | null
    } | null
  } | null
}

export type GetFightsByIdQueryVariables = Exact<{
  reportID: Scalars["String"]["input"]
  fightIDs:
    | Array<InputMaybe<Scalars["Int"]["input"]>>
    | InputMaybe<Scalars["Int"]["input"]>
}>;

export interface GetFightsByIdQuery {
  __typename?: "Query"
  reportData?: {
    __typename?: "ReportData"
    report?: {
      __typename?: "Report"
      title: string
      startTime: number
      endTime: number
      region?: { __typename?: "Region", slug: string } | null
      fights?: Array<{
        __typename?: "ReportFight"
        id: number
        startTime: number
        endTime: number
        difficulty?: number | null
        encounterID: number
        friendlyPlayers?: Array<number | null> | null
      } | null> | null
    } | null
  } | null
}

export type GetCombatantInfoEventsQueryVariables = Exact<{
  reportID: Scalars["String"]["input"]
  fightIDs:
    | Array<InputMaybe<Scalars["Int"]["input"]>>
    | InputMaybe<Scalars["Int"]["input"]>
}>;

export interface GetCombatantInfoEventsQuery {
  __typename?: "Query"
  reportData?: {
    __typename?: "ReportData"
    report?: {
      __typename?: "Report"
      title: string
      startTime: number
      endTime: number
      region?: { __typename?: "Region", slug: string } | null
      events?: {
        __typename?: "ReportEventPaginator"
        data?: any | null
        nextPageTimestamp?: number | null
      } | null
    } | null
  } | null
}
// ...
export interface GetPhysicalDamageTakenEventsQuery {
  __typename?: "Query"
  reportData?: {
    __typename?: "ReportData"
    report?: {
      __typename?: "Report"
      events?: {
        __typename?: "ReportEventPaginator"
        data?: any | null
        nextPageTimestamp?: number | null
      } | null
    } | null
  } | null
}
// ...
export const GetFightsByIdDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "getFightsById" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "reportID" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "fightIDs" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "ListType",
              type: { kind: "NamedType", name: { kind: "Name", value: "Int" } },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "reportData" },
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                {
                  kind: "Field",
                  name: { kind: "Name", value: "report" },
                  arguments: [
                    {
                      kind: "Argument",
                      name: { kind: "Name", value: "code" },
                      value: {
                        kind: "Variable",
                        name: { kind: "Name", value: "reportID" },
                      },
                    },
                  ],
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      { kind: "Field", name: { kind: "Name", value: "title" } },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "startTime" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "endTime" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "region" },
                        selectionSet: {
                          kind: "SelectionSet",
                          selections: [
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "slug" },
                            },
                          ],
                        },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "fights" },
                        arguments: [
                          {
                            kind: "Argument",
                            name: { kind: "Name", value: "translate" },
                            value: { kind: "BooleanValue", value: true },
                          },
                          {
                            kind: "Argument",
                            name: { kind: "Name", value: "fightIDs" },
                            value: {
                              kind: "Variable",
                              name: { kind: "Name", value: "fightIDs" },
                            },
                          },
                        ],
                        selectionSet: {
                          kind: "SelectionSet",
                          selections: [
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "id" },
                            },
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "startTime" },
                            },
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "endTime" },
                            },
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "difficulty" },
                            },
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "encounterID" },
                            },
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "friendlyPlayers" },
                            },
                          ],
                        },
                      },
                    ],
                  },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<GetFightsByIdQuery, GetFightsByIdQueryVariables>;
export const GetCombatantInfoEventsDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "getCombatantInfoEvents" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "reportID" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "fightIDs" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "ListType",
              type: { kind: "NamedType", name: { kind: "Name", value: "Int" } },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "reportData" },
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                {
                  kind: "Field",
                  name: { kind: "Name", value: "report" },
                  arguments: [
                    {
                      kind: "Argument",
                      name: { kind: "Name", value: "code" },
                      value: {
                        kind: "Variable",
                        name: { kind: "Name", value: "reportID" },
                      },
                    },
                  ],
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      { kind: "Field", name: { kind: "Name", value: "title" } },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "startTime" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "endTime" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "region" },
                        selectionSet: {
                          kind: "SelectionSet",
                          selections: [
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "slug" },
                            },
                          ],
                        },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "events" },
                        arguments: [
                          {
                            kind: "Argument",
                            name: { kind: "Name", value: "dataType" },
                            value: {
                              kind: "EnumValue",
                              value: "CombatantInfo",
                            },
                          },
                          {
                            kind: "Argument",
                            name: { kind: "Name", value: "fightIDs" },
                            value: {
                              kind: "Variable",
                              name: { kind: "Name", value: "fightIDs" },
                            },
                          },
                        ],
                        selectionSet: {
                          kind: "SelectionSet",
                          selections: [
                            {
                              kind: "Field",
                              name: { kind: "Name", value: "data" },
                            },
                            {
                              kind: "Field",
                              name: {
                                kind: "Name",
                                value: "nextPageTimestamp",
                              },
                            },
                          ],
                        },
                      },
                    ],
                  },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<